refactor: removed ArkgInstance
