+++ etl/models/__init__.py
@@ -3,7 +3,6 @@ from .anti_recommendation import AntiRecommendation as AntiRecommendation
 from .anti_recommendation_graph_tuple import (
     AntiRecommendationGraphTuple as AntiRecommendationGraphTuple,
 )
-from .arkg_instance import ArkgInstance as ArkgInstance
 from .record_tuple import RecordTuple as RecordTuple
 from .wikipedia_base_url import WIKIPEDIA_BASE_URL as WIKIPEDIA_BASE_URL
