+++ etl/namespaces/arkg.py
@@ -1,5 +1,7 @@
 from pyoxigraph import NamedNode
 
+from etl.models.types import RecordKey
+
 
 class ARKG:
     """A class containing RDF Nodes in the default ARKG namespace."""
@@ -7,3 +9,10 @@ class ARKG:
     BASE_IRI = NamedNode(
         "http://imlapps.github.io/anti-recommender/anti-recommendation/"
     )
+
+    @staticmethod
+    def anti_recommendation_iri(record_key: RecordKey) -> NamedNode:
+        """
+        Return an AntiRecommendation NamedNode for an ARKG.
+        """
+        return NamedNode(ARKG.BASE_IRI.value + record_key)
